# Conflicts:
#	sources/engine/SiliconStudio.Xenko.Assets/Entities/EntityHierarchyCompilerBase.cs
#	sources/engine/SiliconStudio.Xenko.Assets/Entities/SceneAssetCompiler.cs
#	sources/engine/SiliconStudio.Xenko.Engine.Tests/TestEntity.cs
#	sources/engine/SiliconStudio.Xenko.Engine/Engine/EntityManager.cs
#	sources/engine/SiliconStudio.Xenko.Engine/Engine/IEntityComponentNotify.cs

+++ sources/common/core/SiliconStudio.Core.MicroThreading/SchedulerEntry.cs
@@ -9,7 +9,7 @@ namespace SiliconStudio.Core.MicroThreading
     {
         public Action Action;
         public MicroThread MicroThread;
-        public int Priority;
+        public long Priority;
         public long SchedulerCounter;
         public object Token;
 
@@ -18,7 +18,7 @@ namespace SiliconStudio.Core.MicroThreading
             MicroThread = microThread;
         }
 
-        public SchedulerEntry(Action action, int priority) : this()
+        public SchedulerEntry(Action action, long priority) : this()
         {
             Action = action;
             Priority = priority;
